api 폴더 분리, render 메소드 수정
